+++ auth-server/src/main/java/com/truve/platform/user/service/security/properties/KakaoOAuthProperties.java
@@ -1,0 +1,4 @@
+package com.truve.platform.user.service.security.properties;
+
+public class KakaoOAuthProperties {
+}
