+++ auth-server/src/main/java/com/truve/platform/user/service/service/KakaoOAuthService.java
@@ -1,0 +1,4 @@
+package com.truve.platform.user.service.service;
+
+public class KakaoOAuthService {
+}
